requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.5
pinecone-client>=2.2.2
openai>=1.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.11
pandas>=1.5.3
tiktoken>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.26.0
tqdm>=4.66.1